Private Method and JSONRPC names are no longer private
Following names were treated as private names, but these are to be used
in the module that `import`s this module, so making them available to
that module makes sense. To do this these names are no longer private.
May be a good idea at some time to make these as separate `constants`
module.
 - OS_AGENT_KEEPALIVE_MESSAGE
 - OS_AGENT_INIT_MESSAGE
 - JSON_RPC_VERSION_2