`agentUUID` is a required Param
`agentUUID` is a required parameter in both init and keepalive messages.
The 'server' is expected to 'identify' a resource using this 'UUID', if
the resource cannot be found, a 404 needs to be returned. If the
resource is found, all actions are to be taken in the context of this
resource. Later on we might add this 'resource' specific certificates.