Updates to schema
`method` attribute is mandatory only for `*Request` messages, but not
for the response messages. In fact if at all, we'll make it as a part of
params. Moved around that a bit.

Added `method_to_schema_classes` dictionary, which our 'clients' should
use so that this part of the code does not become duplicated.